### 2019-12-21 #1972 @labarba
👋 @majensen — I realize you're handling four submissions now, which is more than the majority of editors have assigned, but I believe this particular submission would be a good fit for you. Can you have a look and tell me if you can take it?
We understand of course that these days will be slow and you might be on holiday. Let us know, and we can assign you for handling this after the New Year.
