### 2019-12-24 #1972 @labarba
Thanks, @majensen! I will assign you now (so the submission comes out of our "without an editor" weekly email), and you can come back to it when you're ready. Everyone is taking a break now, which is good! 🎄
