### 2020-01-08 #1972 @majensen
@gsaksena - would you be willing to review this software submission (#1972) for [JOSS](https://joss.theoj.org)? Details on the review process are [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). Really appreciate your consideration! --@majensen
