### 2020-01-08 #1972 @majensen
Hey @wrathematics, @adavidzh - would either of you be willing to review this submission (#1972) for [JOSS](https://joss.theoj.org]? Thanks for your consideration! --@majensen
