### 2020-01-08 #1972 @adavidzh
> Hey @wrathematics, @adavidzh - would either of you be willing to review this submission (#1972) for [JOSS]([https://joss.theoj.org]](https://joss.theoj.org%5D)? Thanks for your consideration! --@majensen

Thank you for considering me but at this time I am afraid I cannot take on this assignment; while I think I understand the mechanism used this is a topic and an area I am not quite that familiar with.
