### 2020-01-10 #1972 @majensen
Hello @tgerke - would you be able to review this submission (#1972) for [JOSS](https://joss.theoj.org)? Much obliged for your consideration! --@majensen
