### 2020-01-17 #1972 @majensen
Hey @jeremiahsavage - would you be able to review this submission (#1972) for [JOSS](https://joss.theoj.org)? Let me know if you have any questions about the process-Much obliged for your consideration! --@majensen
