### 2020-01-17 #1972 @jeremiahsavage
> Hey @jeremiahsavage - would you be able to review this submission (#1972) for [JOSS](https://joss.theoj.org)? Let me know if you have any questions about the process-Much obliged for your consideration! --@majensen

Hi @majensen , Thanks for thinking of me on this. It looks like interesting an interesting method, but this is out of my expertise.
