### 2020-01-19 #1972 @majensen
@SiminaB - Would you be available to review this submission (#1972) for [JOSS](https://joss.theoj.org)? Let me know if you have questions about the process. I really appreciate your consideration! --@majensen
