### 2020-01-27 #1972 @scwest
@majensen
<https://urldefense.proofpoint.com/v2/url?u=https-3A__github.com_majensen&d=DwMFaQ&c=Cu5g146wZdoqVuKpTNsYHeFX_rg6kWhlkLF8Eft-wwo&r=u6KSyX1ix4-0aQ2-Xzd9a_xhs3muWm7qOpqtGf4LIc8&m=iDPuoPQZE2mHeNc65ocLmSMVb-R7ul0g1j8b7uZ9Hs8&s=Gt106W2CglZorw9L6_YUI_0o6jm9tqN4aCma__hpQ04&e=>
- I understand the difficulty. I can recommend @rhagenson
<https://github.com/RHagenson>. I'll have to review other JOSS editors
online and get back to you.

On Wed, Jan 22, 2020 at 8:54 PM Mark Jensen <notifications@github.com>
wrote:

> @scwest
> <https://urldefense.proofpoint.com/v2/url?u=https-3A__github.com_scwest&d=DwMCaQ&c=Cu5g146wZdoqVuKpTNsYHeFX_rg6kWhlkLF8Eft-wwo&r=u6KSyX1ix4-0aQ2-Xzd9a_xhs3muWm7qOpqtGf4LIc8&m=4k-DUwrefdRgRs5G4pbFUusYFmae38upq1IF153OAik&s=ITA4mMEKQfHLyU2NsiVWrPoGWNFk1tRfLKd1t3_VwNo&e=>
> - having some trouble getting traction on this work. Can you provide
> recommendations for reviewers? I can (prob will) review myself, but want to
> get your input please.
>
