### 2020-01-27 #1972 @rhagenson
@majensen, I am happy to review this.

@scwest, the list of potential reviewers is found [here](https://bit.ly/joss-reviewers) if more reviewers are necessary.
