### 2020-01-27 #1972 @SiminaB
Hi! So sorry, I just saw this! I can try to review it but will need some time (like 2-3 weeks). I also haven't done a JOSS review before now so will need to familiarize myself.
