### 2020-01-27 #1972 @majensen
@SiminaB that would be great! I would be happy to help, and I think the [docs](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) are very good.
