### 2020-01-27 #1972 @majensen
@rhagenson  thanks very much! I believe we have a quorum.
