### 2020-01-27 #2044 @whedon
[REVIEW]: NEEP: null empirically estimated p-values for high-throughput molecular survival analysis
**Submitting author:** @scwest (<a href="http://orcid.org/0000-0002-1394-1999">Sean West</a>)
**Repository:** <a href="https://github.com/thecodingdoc/neep/" target ="_blank">https://github.com/thecodingdoc/neep/</a>
**Version:** 1
