### 2020-01-28 #2044 @rhagenson
@majensen, after reading the conflict of interest statement and reviewing the authors of this paper I have notable conflict of interest as the last author on this paper (Dario Ghersi) is my MS thesis advisor from three years ago -- which I did not notice prior to accepting the offer to review. If you will allow me, I still would like to read the work and offer my help to its authors in the form of opening issues/PRs for necessary fixes prior to acceptance, however I do not believe my opinion can be impartially applied toward the final decision to accept/reject this work.
