### 2020-01-29 #2044 @majensen
Thanks @rhagenson - I appreciate your raising this issue. I myself would definitely find it very helpful if you are willing to assist and even comment in the review. Since we have 2 other reviewers, JOSS guidelines are met. Onward!
