### 2020-02-01 #2044 @rhagenson
@scwest, I opened an issue with all the edits I see as necessary, in addition to suggested edits that I think improve quality-of-life for either yourself as maintainer or others as users.
