### 2020-02-06 e6746be
Adding correct authorship email to previous commits.
