### 2020-02-06 970457a
Adding Makefile and updating README.md
