### 2020-02-06 268d246
Adding directions for testing.
