+++ paper/paper.bib
@@ -6,7 +6,7 @@
 	pages={15(10): e1007469},
 	doi={https://doi.org/10.1371/journal.pcbi.1007469}
 }
-@article{sehgal15,
+@article{sehgal2015,
 	author={Vasudha Sehgal and Elena G. Seviour and Tyler J. Moss and Gordon B. Mills and Robert Azencott and Prahlad T. Ram},
 	year={2015},
 	title={Robust selection algorithm (RSA) for multi-omic biomarker discovery; integration with functional network analysis to identify miRNA regulated pathways in multiple cancers},
