+++ README.md
@@ -6,6 +6,8 @@ Additional processing on NEEP output into protein interaction visualizations can
 
 ## Installation
 
+Installation requires g++ and OpenMP. NEEP can be installed with the following commands.
+
 ```console
 cd ~/neep
 make
@@ -55,4 +57,16 @@ Column 2: best log-rank statistic
 Column 3: null empirically estimated p-value (neep)
 Column 4: adjusted neep
 Column 5: direction (i.e. which survival curve is higher, *low* or *high*)
+
+# Contributing to NEEP
+
+## Making Contributions
+Contributions to this software have been made by Dario Ghersi and Sean West. If you have any desired functionality or would like to contribute to NEEP in some way, please email Sean (seanchristianwest@gmail.com).
+
+## Reporting Issues or Problems with the Software
+If any issues occur with downloading, installation, or running NEEP or if you find issues with the algorithm, please open a GitHub issue in the repository or email Sean (seanchristianwest@gmail.com).
+
+## How to Seek Support
+For questions regarding running or the capabilities of NEEP, add a GitHub issue (https://github.com/thecodingdoc/neep/issues). 
+
 
