+++ dependencies.txt
@@ -1,2 +1,0 @@
-g++
-OpenMP
