### 2020-02-06 ea678d8
Adding JOSS badge.
