### 2020-02-06 fad3681
Adding JOSS badge.
