### 2020-02-06 bfeb716
Removing unused variables (via @wrathematics)
