### 2020-02-06 #2044 @scwest
I have updated the repository with the edit suggestions from @rhagenson. The Makefile and unused variable suggestions from @wrathematics have also been added.
