### 2020-02-18 thecodingdoc/neep#5 @SiminaB
JOSS review #3
My major concerns fall into 3 categories:
1) The summary, statement of need, state of the field, and performance are not very clearly described.
2) Small edits should be made to the README and the test.sh files in order to guarantee a smooth installation and test.
3) More details should be provided for the data example, in terms of the exact processing that should be used or is used by the software.

To detail each of these categories:
1) **The summary, statement of need, state of the field, and performance are not very clearly described.**

- This software presents what appears to be a novel statistical approach. However, it is not entirely what it consists of. The authors mention that a previous approach "could predict the correct p-values" but that "it is not precise enough for p-value correction procedures that are sensitive to very small p-value changes." It is unclear how this was deduced - presumably the Lausen et al approach was tested somewhere and found lacking? Additionally, they mention that the NEEP approach constructs "a null distribution in parallel" (not sure what "in parallel" means) by re-sampling permutations. Thus, it looks like it's a bootstrap-type approach, but this should be clarified.
- The presentation of the state of the field is somewhat confused. For one thing, the log-rank test used for KM approaches is actually the same as the score test applied to a CPH model for a binary variable with no tied survival times (see for example Frank Harrell's comment here https://stats.stackexchange.com/questions/175058/reproducing-the-cox-model-score-test-in-r). For another, the issue with the independent censoring assumption is not that "those who survive to the end of the study are likely to have a higher rate of survival than those who did not" (among other things, studies cannot generally wait until everyone has the event/dies, also it depends when individuals enter the study, as they generally do not enter at the same time), but rather that sicker individuals may be more likely to drop out over the course of the study (see for example https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4282781/). Additionally, it is not clear to me that the NEEP method actually "solves" either the censoring or proportional hazards assumption limitations. Rather, it deals with the issue of obtaining p-values that show good performance despite the selection of the molecular threshold. This should all be explained in more detail and better organized.
- There is no evidence presented that the NEEP approach is "statistically valid," as stated in the summary. Typically this would be done by performing simulations under the null and seeing whether the type I error/FDR are controlled.

2) **Small edits should be made to the README and the test.sh files in order to guarantee a smooth installation and test.**

- I had to use the following steps in order to make the program work on the test cases. These details are not all stated in the README and can make it harder for an inexperienced user. I was able to get it to run using the Ubuntu terminal on Windows 10, so that part is great:
a) Download the entire repository and move it into my home directory.
b) Rename it from "neep-master" to "neep"
c) Open a terminal and change the permissions to allow reading and writing of the directory using sudo chmod -R 775 ~/neep/
d) Edit the code in ~/neep/test/test.sh so that the second line starts with "../neep" rather than "neep". This is because the neep executable is located one level up from test.sh
- I did get small differences between the "test_output.txt" and the "correct_test_output.txt" files. Presumably this is due to the fact that NEEP uses re-sampling permutations?
The first 4 rows/columns of correct_test_output.txt are:

ENST00000310858 | 24.70114 | 71 | 0.00E+00
ENST00000264107 | 23.5003 | 367 | 0.00E+00
ENST00000416789 | 20.87558 | 364 | 2.00E-04
ENST00000409512 | 21.00232 | 367 | 2.00E-04

The first 4 rows/columns of test_output.txt are:

ENST00000310858 | 24.70114 | 71 | 0.00E+00
ENST00000264107 | 23.5003 | 367 | 5.00E-05
ENST00000409512 | 21.00232 | 367 | 1.50E-04
ENST00000416789 | 20.87558 | 364 | 1.50E-04
- The README files states about the output:
"Output file The output will be in tab-separated format with a header. Column 1: molecular object id Column 2: best log-rank statistic Column 3: null empirically estimated p-value (neep) Column 4: adjusted neep Column 5: direction (i.e. which survival curve is higher, low or high)"
However, this is not precise nor is it the exact format of the output file, which includes several other columns.

3) **More details should be provided for the data example, in terms of the exact processing that should be used or is used by the software.**

- The example input file represents an expression matrix. It appears to be of TCGA data, but it unclear what cancer it's from, what technology was used to obtain the data, and what transformations have been applied. For example, it is clear that no log transformation was applied, given that the values are all non-negative and include 0. However, log transformations are generally crucial for this type of analysis. Well-developed pipelines exist for processing microarray, RNAseq, and other types of transcriptomics data.
