### 2020-02-18 #2044 @SiminaB
Apologies for the delay! My initial review is at https://github.com/thecodingdoc/neep/issues/5. I think the paper needs some additional clarifications and details. I also made some suggestions for the README file and some other aspects.
