### 2020-04-03 #2044 @scwest
Hi Mark,

As you might be experiencing yourself, the additional responsibilities
offered by the current COVID-19 pandemic are immediate and pressing.
Specifically as a biomedical researcher, we have lots to do. I should have
reached out earlier, but is it ok if we put this review on hold. Thank you
for your work, I'll be returning to this ASAP. But that looks like it may
be farther out than anticipated.

Sean
