### 2020-05-20 #2044 @majensen
Hi @scwest , checking in with you to see how are things?
