### 2020-05-22 52bafb0
Update paper.md
This addresses the first part of SiminaB's JOSS review.
