### 2020-05-22 thecodingdoc/neep#5 @scwest
Thank you for such a thorough and thoughtful review!

Here are the changes I made to the paper to address this review (changes to the code/testing/documentation to follow):

**The authors mention that a previous approach "could predict the correct p-values" but that "it is not precise enough for p-value correction procedures that are sensitive to very small p-value changes." It is unclear how this was deduced - presumably the Lausen et al approach was tested somewhere and found lacking?**
The original publication gives precision estimates of ~0.001 at its best. For high-throughput data, such as RNAseq with more than 70,000 genes, this is insufficient precision to differentiate the correct order of p-values. Further, in FDR adjustment, small changes to the lowest p-values may drastically alter the number of significant transcripts. To address this confusion, we reworded our sentence to dictate that the precision estimated in the original publication is insufficient for FDR p-value correction.

**Additionally, they mention that the NEEP approach constructs "a null distribution in parallel" (not sure what "in parallel" means) by re-sampling permutations. Thus, it looks like it's a bootstrap-type approach, but this should be clarified.**
The reviewer is absolutely correct! We reworded "in parallel" to "as a bootstrapping procedure implemented in parallel".

**The presentation of the state of the field is somewhat confused. For one thing, the log-rank test used for KM approaches is actually the same as the score test applied to a CPH model for a binary variable with no tied survival times (see for example Frank Harrell's comment here https://stats.stackexchange.com/questions/175058/reproducing-the-cox-model-score-test-in-r).**
We agree with the reviewer. KM is a special case of CPH. In an attempt to simplify the explanation, it seems we misrepresented NEEP! NEEP is not choosing necessarily KM over CPH, it is choosing a binary representation of transcript expression over a continuous one. We address this in the next bit.

**Additionally, it is not clear to me that the NEEP method actually "solves" either the censoring or proportional hazards assumption limitations. Rather, it deals with the issue of obtaining p-values that show good performance despite the selection of the molecular threshold. This should all be explained in more detail and better organized.**
Yes and no. NEEP does not solve censorying and PH directly. Rather, by dealing with the issue of obtaining p-values despite a molecular expression threshold, it allows us to use a binary variable instead of a continuous one. Since molecular expression will follow many patterns of survival-significance, we do not know if the hazard function will be monotonic or proportional. Using a binary variable is certainly not a perfect solution, and as the reviewer states does not solve the proportional hazards assumption. However, a binary variable handles non-monotonic, non-proportional hazard functions better than a continuous variable (see link below). The binary variable response to this is not ideal but we need all the benefit we can get. We address this reviewer comment by restructuring our summary to explicitely discuss binary vs continuous instead of KM vs Cox PH.
(https://bmcmedresmethodol.biomedcentral.com/articles/10.1186/1471-2288-13-88) - see Table 1

**For another, the issue with the independent censoring assumption is not that "those who survive to the end of the study are likely to have a higher rate of survival than those who did not" (among other things, studies cannot generally wait until everyone has the event/dies, also it depends when individuals enter the study, as they generally do not enter at the same time), but rather that sicker individuals may be more likely to drop out over the course of the study (see for example https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4282781/). **
I agree with the second part. We state "rate of survival" to express the probability density function of the lifetime distribution. For studies with an end-date, individuals are censored because of this end-date and not their death. For these end-date censored individuals, their days to death > f(study length, diagnosis date). If we assume that the date of diagnosis is unrelated to the hazard function (which seems reasonable), then we could expect that if given their true days to death (which were censored) they would have a probability density function with a mean survival time larger than those who were not censored. So, I think that what the reviewer states and what we state are saying very similar concepts (if not the same). Ultimately, the logrank power is also dependent on censoring assumptions. So, to address this comment, I removed the statement that "we get around the first assumption by conducting splits" since in retrospect it cannot be validated as true. We also remove discussion of the censorship assumption since it is not directly addressed by NEEP's methodology.

**There is no evidence presented that the NEEP approach is "statistically valid," as stated in the summary. Typically this would be done by performing simulations under the null and seeing whether the type I error/FDR are controlled.**
Because the p-values from NEEP are empirically determined, the type 1 error/FDR are guaranteed to be controlled. For instance, simulations of random molecular expression vectors is equivalent to permuting patients. I.e. the null distribution = random simulations. This is why our method is so precise but also why it is relatively computationally expensive. To address this comment, we explain how the null distribution is equivalent to random expression vectors (p-values are uniform under the null), and that the type 1 error and FDR are guaranteed to be controlled. We remove saying it is "statistically valid" since this is implied after mentioning type 1 error and FDR.
