### 2020-05-22 58c6b95
Response to JOSS review #3, part 2
We alter the test set and the README to address the issues brought up by SiminaB in their associated JOSS review.
