### 2020-05-22 thecodingdoc/neep#5 @scwest
**a) Download the entire repository and move it into my home directory.
b) Rename it from "neep-master" to "neep"**
I added that the installation requires git and added "git clone https://github.com/thecodingdoc/neep.git" as the first command line.

**c) Open a terminal and change the permissions to allow reading and writing of the directory using sudo chmod -R 775 ~/neep/**
Changing the permissions is typically required if downloading from the repositories master instead of using git. So, the "git clone" line added above should handle the permission issue. To handle cases where the user doesn't want to use git, I added to the README:
"If ```git``` is not used but the repository zip is downloaded instead, the directory name should be changed to neep and chmod should be used to allow reading and writing. "

**d) Edit the code in ~/neep/test/test.sh so that the second line starts with "../neep" rather than "neep". This is because the neep executable is located one level up from test.sh**
In the README installation instructions, I specify to add the neep executable to the path variable. In this way, the test.sh can be run even if they remove the executable from the original file. I'm concerned that if the 'mv neep /usr/bin/' in order to put it into the path, that using '../neep' will not work. If they move neep to their path, 'neep' will work no matter what.

**I did get small differences between the "test_output.txt" and the "correct_test_output.txt" files. Presumably this is due to the fact that NEEP uses re-sampling permutations?**
The reviewer is correct! Every time that NEEP is run, the results will be different (with a precision depenent on the size of the null distribution). However, their untied order, the minimum p-values, and the thresholds should never change. Upon further inspection, those two lines had identical adjusted NEEP p-values so their order is not differentiated.
I've also specified in the README the parts of the file that should be the same and which will change. "The output listed in "test\_output.txt" should be nearly identical to that in "correct\_test\_output.txt". The ID, the BEST\_STATISTIC, the BEST\_SPLIT, and the DIRECTION columns should be identical. The remaining columns are subject to changes from the null distribution bootstrapping procedure." I also mentioned that the order of lines with identical adjusted NEEP p-values may be in a different order.

**However, this is not precise nor is it the exact format of the output file, which includes several other columns.**
Another great catch! Since starting the review I had added lines to the output of NEEP which corresponded to survival analysis effect sizes. To address this, I added the missing columns into the README. I also mentioned what the columns were in more detail, and that each line corresponds to one line from the expression matrix input. I also mention that the order of the output is from lowest to highest adjusted NEEP p-value.

**The example input file represents an expression matrix. It appears to be of TCGA data, but it unclear what cancer it's from, what technology was used to obtain the data, and what transformations have been applied. For example, it is clear that no log transformation was applied, given that the values are all non-negative and include 0. However, log transformations are generally crucial for this type of analysis. Well-developed pipelines exist for processing microarray, RNAseq, and other types of transcriptomics data.**
Yes, log transformation are crucial if parametric analysis is performed. However, NEEP uses non-parametric survival analysis. And the relative size of the expression values within an expression vector do not impact the results. In other words, a log transformation doesn't impact the results of NEEP.

Since the data is not used for downstream biological analysis, I don't think details of its generation ought to be included. I think it would be a distraction, since NEEP works on any high-throughput survival analysis data: PPI, gene expression, transcript expression, etc. that the user might want to use. Any method used to preprocess the data that doesn't alter the order of expression across the patients won't effect NEEP. However, you are right to suggest that we give credit to TCGA for the data source. To address this comment, we explicitely mention that as long as patient expression order is not effected, NEEP will give the same results. Further, I give credit to TCGA LUAD as the data source and specify that it has been modified to be a test example. I also give a warning that the test data should not be used for downstream biological analyses.

To address the need to express what kind of data NEEP takes in, I specify that any kind of expression data can be used, giving some examples.
