### 2020-05-22 #2044 @scwest
Hi Mark,

I've submitted a full response to the review #3 (Issue #5). Thank you for
your patience!

Sean

On Wed, May 20, 2020 at 12:22 PM Sean West <seanchristianwest@gmail.com>
wrote:

> I'm ready to unpause this. I'll submit a response to the previous review
> within a couple of days.
>
> Sean
>
> On Wed, May 20, 2020 at 12:17 PM Mark Jensen <notifications@github.com>
> wrote:
>
>> Hi @scwest <https://github.com/scwest> , checking in with you to see how
>> are things?
>>
>> —
>> You are receiving this because you were mentioned.
>> Reply to this email directly, view it on GitHub
>> <https://github.com/openjournals/joss-reviews/issues/2044#issuecomment-631610811>,
>> or unsubscribe
>> <https://github.com/notifications/unsubscribe-auth/AGAU554KI3XJ7HRMMBKX7X3RSQGC7ANCNFSM4KMG7WVA>
>> .
>>
>
