### 2020-06-11 thecodingdoc/neep#5 @SiminaB
Thank you for the revision! Most of my comments have been satisfactorily addressed. I do have a few more notes, some of which are due to my improved understanding of the software/method:

1) My understanding is that the Lausen approach is used to allow for valid inference when choosing the optimal cutoff because using the original p-value is based on making a choice after having seen the data. The statistical reason is the fact that the distribution of p-values under the null is not Uniform(0,1) (you should state that it should be Uniform(0,1), not just uniform, in the paper) but you should provide the intuition that this is because if a researcher picks the cutoff after "looking at the data" (cf Gail and Green https://www.jstor.org/stable/pdf/2529745.pdf, which is cited in the Lausen paper), this introduces a type I error inflation due to essentially "hidden" multiple testing (basically p-hacking) - so instead one has to look at the distribution of the maximum of the test statistic.
2) Essentially your approach speeds up the Lausen method and applies it to many gene/molecular measurements at once, additionally providing the FDR, correct? I think this is much clearer right now, but should be emphasized a bit more, especially in the Statement of Need, which doesn't clearly explain what the software does. Also, for the FDR, you should cite the method you are using (Benjamini-Hochberg?)
3) In the README, it would be good to explain or give a link on how to add ~/neep to the PATH, eg:

> export PATH=$PATH:~/neep

4) The output files have 10 columns and only 9 are described in the README - you forgot the BEST_SPLIT column
5) In the README you state 'The output listed in "test_output.txt" should be nearly identical to that in "correct_test_output.txt". The ID, the BEST_STATISTIC, the BEST_SPLIT, and the DIRECTION columns should be identical. The remaining columns are subject to changes from the null distribution bootstrapping procedure. For rows with tied adjusted NEEP p-values, their order might be different.' Actually, I think all columns except for the p-value and the FDR are identical
between test_output.txt and correct_test_output.txt, since those are the only two affected by the permutations. In the future, it may be helpful to have an option for the seed so the results are truly reproducible.
6) I assume that the -n option under "Usage" in the README specifies the number of permutations. I'm not sure what "size of null distribution to sample" means exactly. This should be clarified.
