### 2020-06-11 #2044 @SiminaB
Alright, I just added my re-review at https://github.com/thecodingdoc/neep/issues/5. I think there are still just a few outstanding issues, but it's much improved!
