### 2020-06-29 thecodingdoc/neep#5 @scwest
Thanks again for your thoughtful review of my response. I have addressed your remaining comments below.

1. **The Lausen approach is used for valid inference when choosing the optimal cutoff because the optimal p-value is chosen after looking at the data; should say Uniform(0,1); type 1 error inflation due to "hidden" multiple testing".**

Yesish. The original reason for this inflation was absolutely because of researchers option of choosing the threshold which fit their needs (one which gave them a significant p-value). However, in the light of >100,000 simultaneous log-rank tests, the problem is no longer a researcher extracting their favorite results after "looking at the data", but instead is the consequence of making an arbitrary threshold choice. If a researcher chooses and sticks to a single threshold, their >100,000 tests are indeed statistically valid. However, the researcher must somehow justify this threshold. NEEP is for those situations where a single threshold choice is unjustifiable (arbitrary); in which case, choosing a single threshold is unacceptably sensitive to the threshold choice (the cited sehgal2015 paper).

To address this comment, I replaced mentions of uniform with Uniform(0,1). I added a sentence that specifies that a single threshold choice cannot be justified for molecular survival analysis based on experimental design. This is in-place of a sentence which provides intuition of a researcher choosing a cutoff after looking at the data. We also added a sentence and cited our paper which showed that small changes in the threshold produced proportionally larger changes in the lists of significant log-rank tests. We believe that not having a justification for a specific threshold in combination with the point that choosing a threshold is sensitive to patient re-sampling / produces a proportionally large change in the results, indicates to the reader the need for a method which provides a statistically valid survival analysis approach in this scenario.

2. **Our approach speeeds up the Lausen method and adds FDR correction; should cite FDR method.**

The Lausen paper is trying to provide a solution for constructing a single log-rank test, without needing to construct the null distribution every single time. In fact, they use MC simulations to provide their gold-standard to compare with their Gaussian-based estimation method, which they deemed a conservative estimate. Since then computers have developed enough, where we believe an estimation method is no longer necessary, and the problem that we address is one massively multiple log-rank tests. So, in effect, we are 1) filtering out molecular elements which do not share from a common null distribution, 2) using a nearly brute-force approach (used by Laussen to create their gold-standard) to generate an unbiased estimate of the null distribution, and 3) using that null to correct p-value. Ultimately, we are not using Lausen's method, rather the method they considered to be their gold-standard, which they were trying to avoid since it was not pragmatic to do this consistently across survival analysis research.

To address this comment, we added two sentences to the research purpose in the statement of need to address the reasoning behind why/when NEEP is beneficial, specifically in the context of choosing a threshold. We also added a citation for BH p-value correction. Then, I reworded the problem solved section to directly state how NEEP addresses this problem.

3. **In the README, it would be good to explain or give a link on how to add ~/neep to the PATH**

Good point! To address this comment, I did as suggested.

4. **Forgot to address BEST_SPLIT column**

Woops. I thought I got them all because I started at 0 when reading the file. I've added in the missing description.

5. **Actually, I think all columns except for the p-value and the FDR are identical
between test_output.txt and correct_test_output.txt, since those are the only two affected by the permutations.**

You are absolutely correct. I'll make the changes in the readme to suggest to that all the non-pvalue columns should be identical.

6. **I assume that the -n option under "Usage" in the README specifies the number of permutations. I'm not sure what "size of null distribution to sample" means exactly. This should be clarified.**

Yes, that is correct! Ultimately, the size of the null distribution that you are sampling is the same as the number of permutations. However, I can see how this is needlessly obfuscative. I put "number of bootstrap samples". I'm worried that using "permutations" also be unclear since permuting is canonically an alternative to bootstrapping (In our case, we our bootstrapping permutations since the permutations themselves describe the null).
