### 2020-06-29 #2044 @scwest
I have posted an updated response for the new comments made by  @SiminaB
<https://github.com/SiminaB>. They are posted in thecodingdoc/neep#5
<https://github.com/thecodingdoc/neep/issues/5>.

Sean
