### 2020-06-29 0c59129
Updated code to address new comments
