+++ paper/paper.bib
@@ -22,3 +22,12 @@
 	journal={Biometrics},
 	pages={73-85}
 }
+@article{benjamini1995,
+	author={Yoav Benjamini and Yosef Hochberg},
+	year={1995},
+	title={Controlling the false discovery rate: a practical and powerful approach to multiple testing},
+	journal={Journal of the Royal statistical society: series B (Methodological)},
+	volume={57},
+	number={1},
+	pages={289-300}
+}
