+++ paper/paper.md
@@ -31,16 +31,22 @@ with the Kaplan-Meier (KM) method.
 Both methods depend on the proportional hazards assumption, but using a
 binary variable performs better when this assumption is violated.
 Yet, KM survival analysis requires that the patients are split into two
-groups by a molecular expression threshold. And the choice of a 
+groups by a molecular expression threshold. 
+In many cases, such as molecular survival analysis, a single threshold 
+choice cannot be justified based on experimental design.
+The choice of a 
 threshold when splitting up a continuous variable has been shown to be
-sensitive to patient group re-sampling [@sehgal2015]. 
+sensitive to patient group re-sampling [@sehgal2015]. And in practice, small
+changes in the chosen threshold produce proportionally 
+larger differences in the set
+of significant logrank tests [@west2019]. 
 
 To circumvent the need for a molecular expression threshold, the logrank
 test can be calculated across a range of thresholds, producing a range 
 of p-values. Choosing the minimum p-value from this range identifies the
 optimal split of patients into two groups, given a single molecular expression
 vector. 
-However, taking the lowest p-value from a range will produce a non-uniform, 
+However, taking the lowest p-value from a range will produce a non-Uniform(0,1), 
 right-skewed distribution of p-values. Since p-values should be uniform
 under the null distribution, the skewed distribution cannot be used for
 valid statistical analysis.
@@ -61,11 +67,12 @@ for this list of molecular expression vector p-values.
 In this way, the precision of the procedure is dependent on the number of 
 samples (permutations) used to generate the null distribution.
 Since the null distribution is generated from the same set of patients,
-the corrected p-values are guaranteed to be uniform under the null if random.
+the corrected p-values are guaranteed to be Uniform(0,1) under the null if random.
 In other words, the minimum p-values for a set of random molecular expression 
 vectors is the same as the null distribution.
 Because of this, the type 1 error and the FDR are guaranteed to be controlled.
-Finally, NEEP conducts False Discovery Rate (FDR) p-value correction and calculates 
+Finally, NEEP conducts False Discovery Rate (FDR) p-value correction 
+[@benjamini1995] and calculates 
 effect sizes, the hazard ratio and the 1, 2, and 5 year mortality
 ratios.
 
@@ -73,9 +80,11 @@ ratios.
 
 # Statement of Need 
 
-**Research purpose**: `NEEP` offers non-parametric, high-throughput, and statistically valid survival analysis of molecular expression vectors.
+**Research purpose**: `NEEP` offers non-parametric, high-throughput, and statistically valid survival analysis of molecular expression vectors. In molecular expression, the choice of a single threshold to separate patients into low- and high-expression cannot be justified and produces variable results.   
 
-**Problem solved**: `NEEP` permutes different orders of patients and calculates their p-values across a range in order to produce a `Null Empirically Estimated P-value' for each molecular expression vector, thus overcoming the assumptions of CPH survival analyiss and the issues of correct p-value estimation. 
+**Problem solved**: `NEEP` chooses the threshold which maximizes the relationship between molecular expression and patient survival for each molecular expression vector. Then `NEEP` corrects the resulting biased p-value distribution using a empirically determined null distribution. 
+To generate the null, `NEEP` permutes different orders of patients and calculates their p-values across a range in order to produce a `Null Empirically Estimated P-value' for each molecular expression vector.
+`NEEP` produces p-values which are uniform under their null distribution so that their precision is dependent on the size of the null generated. By doing so, `NEEP` circumvents the issue of choosing a single threshold while addressing the issue of asymmetric p-value when optimizing the relationship between molecular expression and patient survival.  
 
[…]
