### 2020-07-11 #2044 @majensen
@scwest - re Installation. Need to remove the object files (.o) and the executable (neep) from the repo. Two things: (1) if they are present, then `make` thinks the software is built and doesn't do anything, and (2) they're architecture-dependent and will be useless to many users anyway. Also, please link out to https://www.openmp.org/resources/openmp-compilers-tools/ when mentioning OpenMP as a requirement. I think it would be better to explicitly say "you need a C compiler that complies with OpenMP." Otherwise, it sounds like a library one requires. I will probably help with a PR for your Makefile.
