### 2020-07-11 #2044 @majensen
@scwest - re https://github.com/openjournals/joss-reviews/issues/2044#issuecomment-657073238, I will fork the repo and make a PR with some suggested changes.
