### 2020-07-11 787b474
Remove the make-generated files from repo. Perform a rationally devised automatic test in test.sh
