### 2020-07-11 cba34e0
Update README to reflect new test
