### 2020-07-11 56b23e6
Add OpenMP check to Makefile; update README
