+++ Makefile
@@ -19,6 +19,6 @@ neep: $(OBJS)
 clean:
 	rm -rf *.o a.out check neep
 check:
-	@if $(CXX) $(OMPFLAGS) _.c ; then echo success ;  exit 0 ; else echo "Your compiler does not support OpenMP." ; exit 1 ; fi
+	@if $(CXX) $(OMPFLAGS) _.cpp ; then echo success ;  exit 0 ; else echo "Your compiler does not support OpenMP." ; exit 1 ; fi
 	@touch check
 	
