### 2020-07-21 #2044 @majensen
Hi @scwest - I believe we are in the home stretch. Please review the PR at thecodingdoc/neep#6. I will be looking over the paper itself. thanks!
