### 2020-07-25 #2044 @majensen
@scwest - I am proposing https://github.com/thecodingdoc/neep/pull/7 with some editorial changes.
If you merge both https://github.com/thecodingdoc/neep/pull/6 and https://github.com/thecodingdoc/neep/pull/7, my review will be complete and we can move ahead towards recommendation for acceptance.
thanks
