### 2020-08-19 #2044 @scwest
Thank you Mark for those changes! You really made my life super easy here. I've tested the current master to ensure that it is working and installs correctly on my end. I have generated a new release.
