### 2020-08-21 #2044 @majensen
Hey @scwest - I felt that (1) this work has been in the hopper a long time, (2) if I am a reviewer I should do something, and (3) it was fun and not onerous. I have checked off the remaining items and so both @SiminaB and I now recommend acceptance.

To move forward, can you create a public archive of the latest version of the master branch in a service such as [Zenodo](https://zenodo.org)? Here is a [simple tutorial](https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729) from a previous review.

Once you have done this, can you report back here:
* the new archive's DOI
* the new archive's version number

thanks!
