### 2020-08-26 #2044 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
