### 2020-08-26 #2044 @majensen
@scwest - can you look at https://github.com/openjournals/joss-reviews/issues/2044#issuecomment-680887845 and add those refs? Re: the invalid DOI, I believe you want to reference just the "10.1371/..." portion in the bib file, and whedon adds the prefix. Thanks!
