### 2020-08-26 4538305
Updating paper.bib with proper doi's
For #2044 (comment)
