+++ paper/paper.bib
@@ -4,7 +4,7 @@
 	title={Uncovering and characterizing splice variants associated with survival in lung cancer patients},
 	journal={PLoS Comput Biol},
 	pages={15(10): e1007469},
-	doi={https://doi.org/10.1371/journal.pcbi.1007469}
+	doi={10.1371/journal.pcbi.1007469}
 }
 @article{sehgal2015,
 	author={Vasudha Sehgal and Elena G. Seviour and Tyler J. Moss and Gordon B. Mills and Robert Azencott and Prahlad T. Ram},
@@ -13,14 +13,16 @@
 	journal={PloS one},
 	volume={10},
 	number={10},
-	pages={e0140072}
+	pages={e0140072},
+	doi={10.1371/journal.pone.0140072}
 }
 @article{lausen1992,
 	author={Berthold Lausen and Martin Schumacher},
 	year={1992},
 	title={Maximally selected rank statistics},
 	journal={Biometrics},
-	pages={73-85}
+	pages={73-85},
+	doi={10.2307/2532740}
 }
 @article{benjamini1995,
 	author={Yoav Benjamini and Yosef Hochberg},
@@ -29,5 +31,6 @@
 	journal={Journal of the Royal statistical society: series B (Methodological)},
 	volume={57},
 	number={1},
-	pages={289-300}
+	pages={289-300},
+	doi={10.1111/j.2517-6161.1995.tb02031.x}
 }
