### 2020-08-26 #2044 @scwest
I have added the updated DOIs to the paper.bib file. I then prepared a new release with the changes.
